Remove plugins that fail stricter type checking.
This is another batch of plugins that fail the new type checker. Most of
these failures are either improper array returns, or function signature
mismatches on callbacks. The remainder use old syntax and had improper
int/float coercion or something of that ilk.

Many of these plugins have actual bugs or undefined behavior that is
only now being surfaced.